requirements.txt
# Web server and framework
Flask
gunicorn

# Data processing and optimization
numpy
scipy
pandas
scikit-optimize
SQSnobFit
nevergrad

# JWT token validation
PyJWT[crypto]

# Testing and plotting
matplotlib
